moved file validation to filehandling package
