+++ src/main/java/com/spartaglobal/moviesapi/filehandling/validation/CountryValidationRule.java
@@ -1,0 +1,13 @@
+package com.spartaglobal.moviesapi.filehandling.validation;
+
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateException;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateMoviesException.InvalidCountryException;
+
+public class CountryValidationRule implements FilmValidationRule {
+
+  public void validate(FilmCsvRow film) throws ValidateException {
+    if (film.getCountry() == null || film.getCountry().length() < 2) {
+      throw new InvalidCountryException();
+    }
+  }
+}
