+++ src/main/java/com/spartaglobal/moviesapi/filehandling/validation/CsvRowValidator.java
@@ -1,6 +1,6 @@
-package com.spartaglobal.moviesapi.validation;
+package com.spartaglobal.moviesapi.filehandling.validation;
 
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateException;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateException;
 import java.util.ArrayList;
 import java.util.List;
