+++ src/main/java/com/spartaglobal/moviesapi/filehandling/validation/FilmCsvRow.java
@@ -1,4 +1,4 @@
-package com.spartaglobal.moviesapi.validation;
+package com.spartaglobal.moviesapi.filehandling.validation;
 
 import java.util.List;
