+++ src/main/java/com/spartaglobal/moviesapi/filehandling/validation/FilmValidationRule.java
@@ -1,0 +1,9 @@
+package com.spartaglobal.moviesapi.filehandling.validation;
+
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateException;
+
+public interface FilmValidationRule {
+
+  void validate(FilmCsvRow filmCsvRow) throws ValidateException;
+
+}
