+++ src/main/java/com/spartaglobal/moviesapi/filehandling/validation/GenreValidationRule.java
@@ -1,7 +1,7 @@
-package com.spartaglobal.moviesapi.validation;
+package com.spartaglobal.moviesapi.filehandling.validation;
 
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateException;
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidGenreException;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateException;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateMoviesException.InvalidGenreException;
 import com.spartaglobal.moviesapi.model.enums.Genre;
 
 public class GenreValidationRule implements FilmValidationRule {
