+++ src/main/java/com/spartaglobal/moviesapi/filehandling/validation/TitleValidationRule.java
@@ -1,0 +1,13 @@
+package com.spartaglobal.moviesapi.filehandling.validation;
+
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateException;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateMoviesException.InvalidTitleException;
+
+public class TitleValidationRule implements FilmValidationRule {
+
+  public void validate(FilmCsvRow film) throws ValidateException {
+    if (film.getTitle().length() <= 0) {
+      throw new InvalidTitleException();
+    }
+  }
+}
