+++ src/main/java/com/spartaglobal/moviesapi/filehandling/validation/exceptions/ValidateMoviesException.java
@@ -1,4 +1,4 @@
-package com.spartaglobal.moviesapi.validation.exceptions;
+package com.spartaglobal.moviesapi.filehandling.validation.exceptions;
 
 public class ValidateMoviesException {
