--- src/main/java/com/spartaglobal/moviesapi/validation/CountryValidationRule.java
@@ -1,13 +1,0 @@
-package com.spartaglobal.moviesapi.validation;
-
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateException;
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidCountryException;
-
-public class CountryValidationRule implements FilmValidationRule {
-
-  public void validate(FilmCsvRow film) throws ValidateException {
-    if (film.getCountry() == null || film.getCountry().length() < 2) {
-      throw new InvalidCountryException();
-    }
-  }
-}
