--- src/main/java/com/spartaglobal/moviesapi/validation/FilmValidationRule.java
@@ -1,9 +1,0 @@
-package com.spartaglobal.moviesapi.validation;
-
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateException;
-
-public interface FilmValidationRule {
-
-  void validate(FilmCsvRow filmCsvRow) throws ValidateException;
-
-}
