+++ src/test/java/com/spartaglobal/moviesapi/controller/CsvRowValidatorTest.java
@@ -3,30 +3,30 @@ package com.spartaglobal.moviesapi.controller;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidBudgetException;
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidCountryException;
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidDurationException;
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidGenreException;
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidGross;
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidLanguageException;
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidNameException;
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidRatingException;
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidScoreException;
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidTitleException;
-import com.spartaglobal.moviesapi.validation.exceptions.ValidateMoviesException.InvalidYearException;
-import com.spartaglobal.moviesapi.validation.ActorsNamesValidationRule;
-import com.spartaglobal.moviesapi.validation.BudgetValidationRule;
-import com.spartaglobal.moviesapi.validation.CountryValidationRule;
-import com.spartaglobal.moviesapi.validation.CsvRowValidator;
-import com.spartaglobal.moviesapi.validation.DurationValidationRule;
-import com.spartaglobal.moviesapi.validation.FilmCsvRow;
-import com.spartaglobal.moviesapi.validation.GenreValidationRule;
-import com.spartaglobal.moviesapi.validation.GrossValidationRule;
-import com.spartaglobal.moviesapi.validation.LanguageValidationRule;
-import com.spartaglobal.moviesapi.validation.RatingValidationRule;
-import com.spartaglobal.moviesapi.validation.ScoreValidationRule;
-import com.spartaglobal.moviesapi.validation.TitleValidationRule;
-import com.spartaglobal.moviesapi.validation.YearValidationRule;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateMoviesException.InvalidBudgetException;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateMoviesException.InvalidCountryException;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateMoviesException.InvalidDurationException;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateMoviesException.InvalidGenreException;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateMoviesException.InvalidGross;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateMoviesException.InvalidLanguageException;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateMoviesException.InvalidNameException;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateMoviesException.InvalidRatingException;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateMoviesException.InvalidScoreException;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateMoviesException.InvalidTitleException;
+import com.spartaglobal.moviesapi.filehandling.validation.exceptions.ValidateMoviesException.InvalidYearException;
+import com.spartaglobal.moviesapi.filehandling.validation.ActorsNamesValidationRule;
+import com.spartaglobal.moviesapi.filehandling.validation.BudgetValidationRule;
+import com.spartaglobal.moviesapi.filehandling.validation.CountryValidationRule;
+import com.spartaglobal.moviesapi.filehandling.validation.CsvRowValidator;
+import com.spartaglobal.moviesapi.filehandling.validation.DurationValidationRule;
+import com.spartaglobal.moviesapi.filehandling.validation.FilmCsvRow;
+import com.spartaglobal.moviesapi.filehandling.validation.GenreValidationRule;
+import com.spartaglobal.moviesapi.filehandling.validation.GrossValidationRule;
+import com.spartaglobal.moviesapi.filehandling.validation.LanguageValidationRule;
+import com.spartaglobal.moviesapi.filehandling.validation.RatingValidationRule;
+import com.spartaglobal.moviesapi.filehandling.validation.ScoreValidationRule;
+import com.spartaglobal.moviesapi.filehandling.validation.TitleValidationRule;
+import com.spartaglobal.moviesapi.filehandling.validation.YearValidationRule;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
